pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-observability-otel"
version = "0.1.0"
description = "Multi-agent LangGraph LLM sample with OpenTelemetry traces exported to Azure Monitor"
authors = [ { name = "Your Name", email = "you@example.com" } ]
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
keywords = ["llm", "observability", "opentelemetry", "azure", "langgraph", "fastapi"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
dependencies = [
  "fastapi==0.110.0",
  "uvicorn[standard]==0.29.0",
  "langchain==0.2.12",
  "langgraph==0.0.67",
  "openai==1.50.0",
  "opentelemetry-api==1.30.0",
  "opentelemetry-sdk==1.30.0",
  "opentelemetry-semantic-conventions==0.51b0",
  "opentelemetry-instrumentation-fastapi==0.51b0",
  "opentelemetry-instrumentation-asgi==0.51b0",
  "azure-monitor-opentelemetry-exporter==1.0.0b25",
  "azure-identity==1.18.0"
]

[project.optional-dependencies]
dev = [
  "pytest==8.2.0",
  "coverage==7.6.1",
  "pytest-cov==5.0.0",
  "pytest-asyncio==0.23.6",
  "ruff==0.6.2"
]

[tool.pytest.ini_options]
# Use the single root test directory
pythonpath = ["."]
testpaths = ["test"]
addopts = "-q"

[tool.setuptools.packages.find]
where = ["src"]
include = ["src*"]

[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
show_missing = true
skip_covered = true

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = ["infrastructure"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"
